feat(#48): 잼 박스 다중 삭제 API 개발
- 선택된 잼박스에 속한 링크 삭제
- 잼박스 삭제

+++ src/main/java/com/linkgem/domain/gembox/GemBoxServiceImpl.java
@@ -101,13 +101,19 @@ public class GemBoxServiceImpl implements GemBoxService {
 
     @Transactional
     @Override
-    public void delete(GemBoxCommand.Delete command) {
+    public void deleteGemboxes(GemBoxCommand.Delete command) {
 
-        GemBox gemBox = gemBoxReader.find(command.getId(), command.getUserId())
-            .orElseThrow(() -> new BusinessException(ErrorCode.GEMBOX_NOT_FOUND));
+        List<GemBox> gemBoxList = command.getIds()
+            .stream()
+            .map(gemBoxId -> gemBoxReader.get(gemBoxId, command.getUserId()))
+            .collect(Collectors.toList());
 
-        linkDeleteService.deleteAllByGemBoxId(gemBox.getId());
-        gemBoxStore.delete(gemBox);
+        if(gemBoxList != null && !gemBoxList.isEmpty()) {
+            for(GemBox gemBox : gemBoxList) {
+                linkDeleteService.deleteAllByGemBoxId(gemBox.getId());
+                gemBoxStore.delete(gemBox);
+            }
+        }
     }
 
     @Override
